Implemented MVC
