+++ Frontend/lib/src/views/loginview.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class LoginScreen extends StatelessWidget {
+  const LoginScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: const [
+        Text('TBD'),
+      ],
+    );
+  }
+}
