+++ Frontend/pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   permission_handler: ^10.2.0
+  mvc_pattern: ^8.12.0
 
 dev_dependencies:
   flutter_test:
